Cuts music so that it doesn't continue playing several minutes after the video has ended

+++ ArtlistFootageScraper/Services/FileService.cs
@@ -39,7 +39,10 @@ namespace ArtlistFootageScraper.Services
             string newFilePath = Path.Combine(directory, snakeCaseFileName + extension);
 
             // Rename the file by moving it to the new path
-            File.Move(filePath, newFilePath);
+            if (!File.Exists(newFilePath))
+            {
+                File.Move(filePath, newFilePath);
+            }
 
             return newFilePath;
         }
